Refactored

+++ lib/dtw/shared.py
@@ -1,7 +1,7 @@
 import numpy as np
-from numba import njit, float32  # type: ignore
+from numba import njit, float64  # type: ignore
 
 
-@njit(float32(float32[:], float32[:]), cache=True)
-def cost(a: np.ndarray, b: np.ndarray) -> np.float32:
+@njit(float64(float64[:], float64[:]), cache=True)
+def cost(a: np.ndarray, b: np.ndarray) -> np.float64:
     return np.sum(np.abs(a - b))
